clang-specific Windows delayloading for vulkan

+++ scripts.d/47-vulkan/45-vulkan-loader.sh
@@ -18,12 +18,16 @@ ffbuild_dockerbuild() {
     make install
 
     if [[ $TARGET == win* ]]; then
-        rm "$FFBUILD_PREFIX"/lib/libvulkan-1.dll.a
-        "$GENDEF" "$FFBUILD_PREFIX"/bin/vulkan-1.dll
-        "$DLLTOOL" -d vulkan-1.def --output-delaylib "$FFBUILD_PREFIX"/lib/libvulkan-1.a
-        rm "$FFBUILD_PREFIX"/bin/vulkan-1.dll
+        if [[ $CC == *clang* ]]; then
+            echo 'Libs.private: -Wl,-delayload,vulkan-1.dll' >> "$FFBUILD_PREFIX"/lib/pkgconfig/vulkan.pc
+        else
+            rm "$FFBUILD_PREFIX"/lib/libvulkan-1.dll.a
+            "$GENDEF" "$FFBUILD_PREFIX"/bin/vulkan-1.dll
+            "$DLLTOOL" -d vulkan-1.def --output-delaylib "$FFBUILD_PREFIX"/lib/libvulkan-1.a
+            rm "$FFBUILD_PREFIX"/bin/vulkan-1.dll
 
-        sed -i -e 's/^\(Libs:\).*$/\1 -L${libdir} -lvulkan-1/' "$FFBUILD_PREFIX"/lib/pkgconfig/vulkan.pc
+            sed -i -e 's/^\(Libs:\).*$/\1 -L${libdir} -lvulkan-1/' "$FFBUILD_PREFIX"/lib/pkgconfig/vulkan.pc
+        fi
     elif [[ $TARGET == linux* ]]; then
         gen-implib "$FFBUILD_PREFIX"/lib/libvulkan{.so.1,.a}
         rm "$FFBUILD_PREFIX"/lib/libvulkan.so*
